fix(sofya): address review feedback
- changeset: add the public API usage example required for feature changesets
- search: enforce integer maxResults and max 10 include/exclude domains
- research: enforce integer maxSources
- docs: remove em dash, set initial package version to 0.0.0

freshness stays a string because Sofya also accepts date ranges (YYYY-MM-DD), which an enum cannot express.

+++ integrations/sofya/src/research.ts
@@ -11,7 +11,7 @@ const inputSchema = z.object({
     .string()
     .optional()
     .describe("Filter sources by recency. One of 'day', 'week', 'month', 'year', or a 'YYYY-MM-DD' date"),
-  maxSources: z.number().min(1).max(30).optional().describe('Maximum number of sources to use (1-30)'),
+  maxSources: z.number().int().min(1).max(30).optional().describe('Maximum number of sources to use (1-30)'),
 });
 
 const outputSchema = z.object({
